Remove openstack.common.db.common
Since this only contains exceptions that are either used in session.py
or utils.py (not both), it seems sensible to move them into the modules
where they are used.

Also, remove the use of the Invalid base exception class since we don't
seem to be making use of the base class anywhere by catching it rather
than the more specialized exceptions.

Change-Id: Ib05bb2e0a9494e1dc60c60b8eee0e76b5d2ee555