Add a fixture for using of SQLite in-memory DB.
Currently, SQLite DB in a file is used for testing of DB related
code. Thus, in each test case we have to ensure that tearDown()
drops all created tables, etc. It is much easier and cleaner to
guarantee that DB is in the original state by creating it in
memory instead.

Change-Id: I6f0a9735a0fb31bef30842afff50c6089fa2bf92