Remove lazy loading of database backend
A single DBAPI instance is created globally in target projects at module
import time. As those projects are mostly API services (i.e. DB
oriented), there is no much sense in lazy loading of a DB backend. On
the other hand, if we initialize the DB backend as soon as a DBAPI
instance is created, we simplify code a lot, get rid of the lock and
remove the overhead of checking if the backend has been initialized from
every __getattr__() call.

Change-Id: I9fdf3eab32ee0f078d64fbde35df81c2d1123ae2