Incorporating MIT licensed code
This patch attempts to clarify the licensing position
held by the db.sqlalchemy.migration module. According to
https://wiki.openstack.org/wiki/LegalIssuesFAQ when we have MIT licensed
code in the project we have to update the toplevel LICENSE file to include
the details of the license + a brief explanation of which code is under
that license.

This patch also fixes the license header on that module which was
incomplete.

Change-Id: I5675f3f6f63ac6432736a98e94fbe2cb3e32714f