Make table utf-8 charset checking be optional for DB migration
Add an option to the migration db_sync call to allow a project to
disable the check for utf8 encoding on all tables.  This allows the
project to fix problems automatically via a migration, which
otherwise would not be possible because the utf8 check would stop
them from running migrations.

Change-Id: I7f271d846141ac72dde3fb0d12159b125018eb2c
Closes-bug: #1279000
Signed-off-by: Zhi Yan Liu <zhiyanl@cn.ibm.com>