Add a warning to not use get_table for working with ForeignKeys
The get_table utility loads a single table with a unique MetaData object
each time it's called but sqlalchemy needs the MetaData object to hold
information about both tables in the ForeignKey constraint so it
shouldn't be used when creating those types of constraints.  This patch
adds a warning to the method's docstring against it's usage in that
case.

Closes-Bug: #1303824

Change-Id: I7b038d668f08f166700300fd787a8cf3cab73a17