Do not always adjust sqlalchemy.engine logging
Instead of messing with another libraries logger by default
only adjust the levels of that logger if the connection debug
is greater than or equal to zero. This allows those who have
configuration files or other ways to adjust logging to not be
affected by this codes level manipulation (which is likely
a side-effect users would not expect).

Change-Id: I8c52f595e5077627c6999deadfd748034ab02fff