Fix changing the type of column deleted
SQLite doesn't provide a native implementation of BOOLEAN data type.
SQLAlchemy emulates BOOLEAN data type for SQLite using INT column +
a CHECK constraint.

We've been providing a helper util to change the type of column
'deleted' from BOOLEAN to INT. Due to limitations of ALTER in SQLite,
in order to do that, we omit the original 'deleted' column as well as
the corresponding CHECK constraint when recreating the table.
Omitting of the constraint is more tricky, though. It's implemented
by analyzing the SQL text used for its rendering.  SQLAlchemy
versions 0.8.3+ slightly changed the way constraints are rendered, so
we have to update our util changing the type of 'deleted' column to
work with SQLAlchemy 0.8.3+ on SQLite backend (we aren't using SQLite
in production, but we still need it for running of unit tests).

Closes-Bug: #1252693

Change-Id: Ie92caff005217250de17461f4f87692ad8e96a09