Check for mysql_sql_mode is not None in create_engine()
The check for mysql_sql_mode uses an empty boolean check,
which misses the case where sql_mode is passed as the blank string.
As a result, the blank string value is not passed to SET SESSION,
causing tests to break where a blank value is expected.
The conditional is changed to "is not None".

Change-Id: I919ab7c10aadf4b8778bd1e3425b9221fcdaa2f8
Closes-Bug: 1340395