Handle DB2 SmallInteger type for change_deleted_column_type_to_boolean
With the DB2 sqlalchemy driver (ibm_db_sa), boolean columns are stored
as SmallInteger rather than Boolean (or TINYINT for mysql). This change
updates the two tests that were changing the deleted column to boolean
and then checking the type based on the backing engine so they work for
ibm_db_sa.

Also changes the assertion from assertTrue to assertIsInstance since it
will provide a clearer failure message.

Closes-Bug: #1312423

Change-Id: I42975119ecacaac89e7225f57f9f6b52c455ab6a