Remove utils.drop_unique_constraint()
This function is currently otherwise unused and does not
work correctly on SQLite, not only dropping the
target unique constraint but instead dropping all unique
constraints from the target table prior to SQLAlchemy
version 1.0 (unreleased).   The tests miss that it isn't
working as well unless SQLAlchemy 1.0 is used, in which
case they fail.

Change-Id: Icaff621fba4df289247e96b87c439c62e91543d6
Closes-bug: #1377646