Organize provisioning to use testresources
This change introduces the use of the testresources package,
such that the provisioning system uses TestResourceManager
objects in order to create and drop databases, schemas,
and manage transactional testing.  A new series of objects
to support transparent transaction containers within
tests is added as well.

partially implement bp: long-lived-transactionalized-db-fixtures
Partial-Bug: #1339206

Change-Id: I16bfa3af0e1ad6a9231ea38dea7cd76092347f55