Implement backend-specific drop_all_objects for provisioning.
This patch implements the drop_all_objects() portion
of the testresources series of changes, so that this
utility is usable right now.  In particular, as SQLAlchemy 1.0
will support reflection of foreign key names in SQLite,
it's necessary that we don't emit a DROP CONSTRAINT
for these against SQLite as SQLite does not support
this command.

The drop_all_objects() is added as a first class element
here and is also added to the _cleanup() method of
the test_migrations base.

Tests are added which exercise the case of mutually-dependent
foreign keys on Postgresql and MySQL as well.

partially implement bp: long-lived-transactionalized-db-fixtures
Partial-Bug: #1339206

Change-Id: I56777834187651f5ba2e43482546c2981524b11c