Sort model fields using getattr(), not inspect()
While inspect(mapping).all_orm_descriptors is a good
way to get at attributes that are known to be mapped or
otherwise described by SQLAlchemy, it does not produce a
SQL expression in the case of a hybrid descriptor, where it
produces the descriptor, not the result of calling it.
This fix ensures that the descriptor is evaulated, while
still retaining the fact that we check the attribute
as coming from an ORM-level descriptor.

Change-Id: I04a03be713348158b8cdfda683624fea01fe34ea
Closes-Bug: #1451880