Fix tests to work under both pymsysql 0.6.2 and 0.7.x
In 0.6.2, the exc.statement and exc.params were as follows:
exc.statement = {unicode} u'INSERT INTO resource_entity (id, foo_id)
                VALUES (%s, %s)'
exc.params = {tuple} (1, 2)

In 0.7.1, the exc.statement and exc.params are as follows:
exc.statement = {unicode} u'INSERT INTO resource_entity (id, foo_id)
                VALUES (%(id)s, %(foo_id)s)'
exc.params = {dict} {'id': 1, 'foo_id': 2}

So the easiest thing to support both is to use string interpolation
and then check if they are equal. This approach works for both py34
and py27 as well. Note that we still have to test the old way for
sqlite which has (?, ?) in the SQL statement.

Closes-Bug: #1533861
Change-Id: Iab9f0e4b4dd337363dd7e705c1155f3f6517b404