Fix db_version checking for sqlalchemy-migrate
When a project has multiple migration repos and
oslo_db.sqlalchemy.migration.db_version() is called using the one
repo as the abs_path, it will attempt to put the repo under version
control. The next time it's called with the other abs_path, it would
fail rather than attempting to put the repo under version control
when using sqlalchemy-migrate. When using alembic, db_version would
attempt to put the new abs_path under version control.

This change makes it so that when sqlalchemy-migrate is used
db_version() will attempt to version control the new abs_path just
like when alembic is used.

Change-Id: If5e2ae1a72f9118c7aca34e925fdc5e3e8480275