Catch empty value DBDuplicate errors
We can have integrity constraint violations when the
value is an empty string. The previous regex for converting
these into DBDuplicate exceptions was not considering this
case.

This adjusts the regex to a 0 or many instead of a 1 or many
match for the value and adds a case to a test.

Change-Id: Ibf4686f92fcffb8e70ec2735ab878e8849ed37b1
Closes-Bug: #1599086
(cherry picked from commit 16886e55bd8fde77f71d75c3c87c2f71ee6dbd4f)