utils: fix get_unique_keys() when model has no table attached to it
Using the internal SQLAlchemy __table__ to find unique keys on a model
class does not work if the model is not attached to a class. In this
case, it seems there is no way to find what the keys are: rather than
crashing or returning garbage, the function just returns None,
indicating it does not know how to get it.

The warning from paginate_query() is not emitted in this case, as
there's no way to guess if the developer did wrong or not.

Change-Id: I1291fe1f1e78471f18f230a66edff31ed199bc7a
(cherry picked from commit 49eaed489e4650e56ed2e270ba3e15ce2b421cc8)