Fix exc_filters for mysql-python
Something changed in mysql-python / libmysqlclient or mysql and now
exceptions that used to be InternalError's have become OperationalError's.
Update filters accordingly to wrap those into oslo.db exceptions
properly.

OperationalError's are kind of special, as we re-raise them as is due
to historical reasons: we need to make sure that at least SAVEPOINT
errors are wrapped, as that's what oslo.db and neutron unit tests
expect.

Closes-Bug: #1630899

Change-Id: I39ea8c4a0dc9d43992f00da7b1ff502595b4dc13