Support MariaDB error 1927
We're observing a MariaDB-specific error code
"1927 connection was killed", which seems to indicate the condition
where the client/server connectivity is still up but the mariadb
server has been restarted.  This error needs to be supported
as one of the "disconnect" situations so that the ping handler
knows to force a reconnect.

Change-Id: I484237c28a83783689bb8484152476de33f20e3a
References: https://bugs.launchpad.net/tripleo/+bug/1642944