Support packet sequence wrong error
In I484237c28a83783689bb8484152476de33f20e3a, we observed
MariaDB-specific error code 1927 as a new kind of disconnect situation.
In some failure-mode testing we are also observing with PyMySQL that
a stale connection, when it is attempted to be used for the initial
ping, does not return any error code but instead trips up
PyMySQL with packets sequence numbers (this is an internal value
used by the MySQL client/server protocol to track messages).
The connection is definitely not recoverable in this state so this also
needs to be a disconnect situation so that the ping listener refreshes
the connection pool.

Change-Id: Iee1c4533e2457acc30e2046b63145c2502012c5f
References: https://bugs.launchpad.net/tripleo/+bug/1642944
(cherry picked from commit c61d2d47bd39c31f1cd06faf497ffa85ae66cdd4)