Establish flush() for "sub" facade contexts
When converting existing session context manager code to
use enginefacade, normally a flush() is implicit on
a "subtransaction" when it closes.  Enginefacade
has not maintained this behavior.  Ideally, this should
have been on by default when enginefacade was first
released, but for now add this as a behavioral
option for projects that require it.

Co-authored-by: Mike Bayer <mike_mp@zzzcomputing.com>
Closes-Bug: #1664643
Change-Id: I6f4f439928588cff954e749dfa938425892e0931