Attach context being used to session/connection info
Attaching a reference to the context to the session/connection
is useful for SQLAlchemy event handlers that only get access to
the session and need access to the context for additional details.

The use case driving this is the related bug where we want to pass
conditional update constraints set in the context via the HTTP API
down into an SQLAlchemy event listener.

Related-Bug: #1493714
Change-Id: I5c08e672782344e8778187e257e03c1a1d1b019a