Warn on URL without a drivername
Older installations may still be running with
URLs of the form "mysql://user:pass@host/dbname",
which defaults to the native MySQL driver that is not supported
by eventlet.  Warn that URLs should be qualified with an explicit
driver, and in the case of MySQL that PyMySQL should be used.

Change-Id: Ie973a43c8d056778d02703cf75bfc52a90027dea