Log an exception when reconnect-to-disconnected occurs
The scenario where an existing database connection that's
present in the pool is found to be not connected to the
database should normally not be occurring; the idle_timeout
setting (aka pool_recycle) should be less than the effective "idle timeout"
set up on the server and/or HAProxy such that a stale
connection in the pool should already have been recycled.

As a system with a mis-configured idle_timeout can experience
latency due to this error hook, it should be reported that
this is an avoidable situation which is occurring. Right
now there's no easy way to see if a performance test is
spending lots of time hitting this condition or not.

Change-Id: I3e9613cbb982c727762247201c5c4972e21da725