Reverse role of synchronous_reader
Repaired the "synchronous_reader" modifier of enginefacade so that it
refers to the "writer" engine when set to True, thereby allowing
"synchronous" behavior with the writer.  When set to False, this is
"asynchronous", so this should be associated with the async engines.
The flag had the reverse behavior previously.

Change-Id: Id7fea7562ba90eb710176d497af103303f230531
Closes-bug: #1746116
(cherry picked from commit 5ca329884d492d78e623e2313142238eb256f1a6)