Remove convert_unicode flag
This flag is to be deprecated in SQLAlchemy version 1.3,
as it is a leftover from the pre-Python 3 legacy days when
DBAPIs under Python 2 would crash if you passed them a
u'' object.   While still available in 1.3, it will emit a
deprecation warning, so it's time to stop using this flag and to
locate any downstream Openstack projects that might be relying
on its behavior in a Python 2 context; the flag has no real effect
on Python 3.

Change-Id: I01984b4ff7e03dcf91ef8dcb05da67dda8270ae3
References: https://github.com/sqlalchemy/sqlalchemy/issues/4393