Support context function argument as keyword
Today, when a user uses enginefacade reader/writer decorators
he/she is forced to pass the context argument as the
first positional argument. performing a decorated function
call with this argument passed as a keyword would lead to
an out of range exception.

This patch proposes to add support for the context argument
to be passed as a keyword argument in decorated functions.

e.g

@enginefacade.reader
def foo(context):
    ...

foo(context=ctxt)  <- will now be possible

Change-Id: Ief0b71bf9a7eb75935612431bdcc26d33bce852d