Remove LegacyEngineFacade
We need to do this rather invasive change at some point. The
LegacyEngineFacade has been deprecated since 1.12.0 way back in 2015.
It's time to finally remove it, along with the helpers that generate
such an object.

The 'OsloDBDeprecationWarning' exception is also removed. This could
possibly do with a deprecation cycle, but a quick check of
codesearch.o.o suggests this was only ever used internally and can
therefore be removed with a minimum amount of noise.

Depends-on: https://review.opendev.org/c/openstack/glance/+/926560

Change-Id: I169aaf279f6abd675154bba6b16d040ae7b61028
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>