tests: Enable SADeprecationWarning warnings
Highlight use of deprecated SQLAlchemy APIs to ensure we keep on top of
things. This requires resolving the following SADeprecationWarning
warnings:

  The from_engine() method on Inspector is deprecated and will be
  removed in a future release.  Please use the sqlalchemy.inspect()
  function on an Engine or Connection in order to acquire an Inspector.

  The Column.copy() method is deprecated and will be removed in a future
  release.

  The ColumnCollectionConstraint.copy() method is deprecated and will be
  removed in a future release.

There are more warnings to be resolved related to SQLAlchemy 2.0, but
those require a special environment option (SQLALCHEMY_WARN_20) to
trigger and a lot of work to resolve. We'll address those in a series of
follow-ups.

Change-Id: I34b395e6d50f4e4151178c327d94308e6f5d5b6e
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>