Ignore unset opts when deprecated
If an option is deprecated, there's a good chance we want to ignore it
unless explicitly requested by the user. Start ignoring options that are
deprecated and unset.

Conflicts:
  oslo_db/tests/sqlalchemy/test_sqlalchemy.py

NOTE(stephenfin): Conflicts are because we still have the autocommit
parameter in our engine calls. This was removed in master in
preparation for SQLAlchemy 2.0.

Change-Id: I3008562db52a5625504b262d7148528a832246dd
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Closes-Bug: #2008720
(cherry picked from commit f6ba070657b3bb6b57cef3728b3e6b95f0bf08cb)