[stable-only] Re-add autocommit support
There are a number of projects that have yet to migrate away from
autocommit. To prevent them blocking the release, temporarily re-add
autocommit support. We do this only on the stable branch since we are
going to need to address SQLAlchemy 2.x support in C.

This is effectively a partial revert of change
Ifaca67c07f008d8bc0febeecd3e200cc7ee7a4b0 but focused solely on
autocommit and with additional logic in tests to prevent us trying to
use autocommit with SQLAlchemy 2.x (where it's not supported).

Change-Id: Ib49ed6886470284e6486581290af4e0f2482e5c5
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>