[stable-only] Re-add migration helpers, tests
There are a number of projects that have yet to migrate from
SQLAlchemy-Migrate to Alembic and are relying on some old
SQLAlchemy-Migrate helpers. To prevent them blocking the release,
re-add the helpers (though not the tests for same since those won't pass
with SQLAlchemy 2.x). We only do this on the stable branches since we
are going to need to address SQLAlchemy 2.x support in C.

This is effectively a partial revert of change
Ic3d6bd318038d723b0d50d39e45f8e26289e9a57 but it only re-adds the test
helper rather than the entirety of oslo.db's former SQLAlchemy-Migrate
support.

Change-Id: Icdb6a83be28e731b3a99df88b20d27829bcf50be
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>