Python3.14: fix tests for AsyncioFacadeTest
In Python3.14, AsyncioFacadeTest tests fail with:

"ValueError: the greenlet library is required to use this function. No
module named 'greenlet'"

It is unclear why it works in Python 3.13, but SQLAlchemy does state
that one should install SQLAlchemy[asyncio] in order to use greenlet:

"The SQLAlchemy asyncio module requires that the Python 'greenlet'
library is installed. In order to ensure this dependency is available,
use the 'sqlalchemy[asyncio]' install target: 'pip install
sqlalchemy[asyncio]'"[1]

[1] https://github.com/sqlalchemy/sqlalchemy/blob/2e058575c237f33058da79433f6f1ce5f6739121/lib/sqlalchemy/util/concurrency.py#L40

Change-Id: Id1eaeca5bc2ce62ff926b0404a012960c73f04d2
Signed-off-by: Cyril Roelandt <cyril@redhat.com>